move commands handling onto commands/ for better codebase structure
